+++ lib/main.dart
@@ -1,125 +1,114 @@
+import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/material.dart';
+import 'package:pactus_gui_widgetbook/app_scaffold.dart';
+import 'package:pactus_gui_widgetbook/src/core/theme/app_theme.dart';
+import 'package:pactus_gui_widgetbook/src/core/theme/app_theme_data.dart';
+import 'package:widgetbook/widgetbook.dart';
+import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
+import 'package:flutter/widgets.dart';
+import 'package:pactus_gui_widgetbook/src/core/utils/gen/localization/app_locales.dart';
+import 'package:pactus_gui_widgetbook/src/core/utils/gen/localization/customized_localization_addon.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'main.directories.g.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(const WidgetBookApp());
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+/// ## [WidgetBookApp] Class Documentation
+///
+/// The `WidgetBookApp` class is the main entry point for
+/// the Widgetbook application.
+/// It defines the Widgetbook configuration, including the devices,
+/// themes, and addons used for visualizing and testing widgets.
+///
+/// ### Usage:
+///
+/// The `WidgetBookApp` sets up a Widgetbook environment with
+/// predefined themes, devices, and alignment options.
+/// It enables developers to test their widgets in different
+/// contexts and layouts.
+///
+/// ### Addons:
+///
+/// - **[DeviceFrameAddon]**:
+///   - Includes a list of devices for testing widgets.
+///   - **Devices**:
+///     - `Devices.ios.iPhone13`: Simulates an iPhone 13 device.
+///     - `Devices.windows.laptop`: Simulates a standard laptop.
+///     - `Devices.windows.wideMonitor`: Simulates a widescreen monitor.
+///
+/// - **[AlignmentAddon]**:
+///   - Allows testing widget alignment within the parent container.
+///
+/// - **[ThemeAddon]**:
+///   - Provides two themes for testing:
+///     - **Light Theme**: Uses `AppThemeData.lightTheme()` for a
+///     bright interface.
+///     - **Dark Theme**: Uses `AppThemeData.darkTheme()` for a
+///     darker interface.
+///   - Includes a custom theme builder that wraps widgets with
+///   `AppTheme` and `AppScaffold`.
+///
+/// ### Properties:
+///
+/// - **[directories]**:
+///   - Dynamically generated list of widget directories.
+///   - Facilitates organizing and testing widgets within Widgetbook.
+///
 
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // TRY THIS: Try running your application with "flutter run". You'll see
-        // the application has a purple toolbar. Then, without quitting the app,
-        // try changing the seedColor in the colorScheme below to Colors.green
-        // and then invoke "hot reload" (save your changes or press the "hot
-        // reload" button in a Flutter-supported IDE, or press "r" if you used
-        // the command line to start the app).
-        //
-        // Notice that the counter didn't reset back to zero; the application
-        // state is not lost during the reload. To reset the state, use hot
-        // restart instead.
-        //
-        // This works for code too, not just values: Most code changes can be
-        // tested with just a hot reload.
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
-      ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
-    });
-  }
+@widgetbook.App()
+class WidgetBookApp extends StatelessWidget {
+  const WidgetBookApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
-    return Scaffold(
-      appBar: AppBar(
-        // TRY THIS: Try changing the color here to a specific color (to
-        // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
-        // change color while the other colors stay the same.
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
-      ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          //
-          // TRY THIS: Invoke "debug painting" (choose the "Toggle Debug Paint"
-          // action in the IDE, or press "p" in the console), to see the
-          // wireframe for each widget.
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
+    return Widgetbook(
+      appBuilder: (context, child) {
+        return child;
+      },
+      addons: [
+        CustomizedLocalizationAddon(
+          locales: AppLocales.supportedLocales,
+          localizationsDelegates: const [
+            AppLocalizations.delegate,
+            GlobalMaterialLocalizations.delegate,
+            GlobalWidgetsLocalizations.delegate,
+            GlobalCupertinoLocalizations.delegate,
+          ],
+          initialLocale: AppLocales.supportedLocales.first,
+        ),
+        DeviceFrameAddon(
+          devices: [
+            Devices.ios.iPhone13,
+            Devices.windows.laptop,
+            Devices.windows.wideMonitor,
           ],
         ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
+        AlignmentAddon(),
+        ThemeAddon(
+          themes: [
+            WidgetbookTheme(
+              name: 'Light',
+              data: AppThemeData.lightTheme(),
+            ),
+            WidgetbookTheme(
+              name: 'Dark',
+              data: AppThemeData.darkTheme(),
+            )
+          ],
+          themeBuilder: (context, data, child) {
+            return AppTheme(
+              themeData: data,
+              child: AppScaffold(child: child),
+            );
+          },
+        ),
+      ],
+      // The [directories] variable does not exist yet,
+      // it will be generated in the next step
+      directories: directories,
     );
   }
 }